+++ src/system/restart.c
@@ -10,6 +10,7 @@
 #include <sys/mman.h>
 #include <sys/types.h>
 #include <sys/wait.h>
+#include <sys/signalfd.h>
 
 #include "hin.h"
 
@@ -91,17 +92,17 @@ int hin_restart () {
   return 0;
 }
 
-static void sig_restart (int signo) {
+static void hin_sig_restart_handler (int signo, siginfo_t * info, void * ucontext) {
   hin_restart ();
 }
 
-static void sig_child_handler (int signo) {
+static void hin_sig_child_handler (int signo, siginfo_t * info, void * ucontext) {
   pid_t pid;
   int status;
 
   while ((pid = waitpid (-1, &status, WNOHANG)) > 0) {
     if (WIFSIGNALED (status)) {
-      if (WTERMSIG(status) == SIGSEGV) {
+      if (WTERMSIG (status) == SIGSEGV) {
         // It was terminated by a segfault
         printf ("child %d sigsegv\n", pid);
       } else {
@@ -124,25 +125,130 @@ static void sig_child_handler (int signo) {
   }
 }
 
-int install_sighandler () {
+static void hin_sig_pipe_handler (int signo, siginfo_t * info, void * ucontext) {
+  printf ("broken pipe signal received\n");
+}
+
+static void hin_sig_int_handler (int signo, siginfo_t * info, void * ucontext) {
+  printf("^C pressed. Shutting down.\n");
+  void hin_clean ();
+  hin_clean ();
+  exit (0);
+}
+
+int hin_signal_clean () {
+  signal(SIGINT, SIG_DFL);
+  signal(SIGUSR1, SIG_DFL);
+  signal(SIGPIPE, SIG_DFL);
+  signal(SIGCHLD, SIG_DFL);
+  return 0;
+}
+
+int hin_signal_install () {
   // It's better to use sigaction() over signal().  You won't run into the
   // issue where BSD signal() acts one way and Linux or SysV acts another.
   struct sigaction sa;
   memset (&sa, 0, sizeof sa);
   sigemptyset (&sa.sa_mask);
-  sa.sa_flags = 0;
+  sa.sa_flags = SA_SIGINFO;
 
-  sa.sa_handler = SIG_IGN;
+  sa.sa_sigaction = hin_sig_int_handler;
+  sigaction (SIGINT, &sa, NULL);
+
+  sa.sa_sigaction = hin_sig_pipe_handler;
   sigaction (SIGPIPE, &sa, NULL);
 
-  sa.sa_handler = sig_restart;
+  sa.sa_sigaction = hin_sig_restart_handler;
   sigaction (SIGUSR1, &sa, NULL);
 
-  sa.sa_handler = sig_child_handler;
+  sa.sa_sigaction = hin_sig_child_handler;
   sigaction (SIGCHLD, &sa, NULL);
 
   return 0;
 }
+
+#if 0
+
+static sigset_t mask;
+static hin_buffer_t * signal_buffer = NULL;
+
+int hin_signal_clean () {
+  if (sigprocmask (SIG_UNBLOCK, &mask, NULL) == -1)
+    perror ("sigprocmask");
+  if (signal_buffer) {
+    close (signal_buffer->fd);
+    hin_buffer_clean (signal_buffer);
+    signal_buffer = NULL;
+  }
+  return 0;
+}
+
+static int hin_signal_callback1 (hin_buffer_t * buf, int ret) {
+  struct signalfd_siginfo * info = (void*)buf->ptr;
+  printf ("got sighandler\n");
+
+  if (ret < 0) {
+    printf ("signal error read '%s'\n", strerror (-ret));
+    hin_request_read (buf); // ?
+    return 0;
+  }
+
+  switch (info->ssi_signo) {
+  case SIGINT:
+    hin_sig_int_handler (info->ssi_signo, info, NULL);
+  break;
+  case SIGUSR1:
+    hin_sig_restart_handler (info->ssi_signo, info, NULL);
+  break;
+  case SIGPIPE:
+    hin_sig_pipe_handler (info->ssi_signo, info, NULL);
+  break;
+  case SIGCHLD:
+    hin_sig_child_handler (info->ssi_signo, info, NULL);
+  break;
+  default:
+    printf ("error got unexpected signal\n");
+    return -1;
+  break;
+  }
+
+  hin_request_read (buf);
+
+  return 0;
+}
+
+int hin_signal_install () {
+  sigemptyset (&mask);
+  sigaddset (&mask, SIGINT);
+  sigaddset (&mask, SIGUSR1);
+  sigaddset (&mask, SIGPIPE);
+  sigaddset (&mask, SIGCHLD);
+
+  if (sigprocmask (SIG_BLOCK, &mask, NULL) < 0)
+    perror ("sigprocmask");
+
+  int sig_fd = signalfd (-1, &mask, 0);
+  if (sig_fd < 0)
+    perror ("signalfd");
+
+  hin_buffer_t * buf = malloc (sizeof (*buf) + sizeof (struct signalfd_siginfo));
+  memset (buf, 0, sizeof (*buf));
+  buf->flags = 0;
+  buf->fd = sig_fd;
+  buf->callback = hin_signal_callback1;
+  buf->count = buf->sz = sizeof (struct signalfd_siginfo);
+  buf->ptr = buf->buffer;
+  buf->debug = 0xffffffff;
+  signal_buffer = buf;
+
+  printf ("installed sighandler fd %d\n", sig_fd);
+
+  hin_request_read (buf);
+
+  return 0;
+}
+
+#endif
 
 static int hin_use_sharedmem (int sharefd) {
   hin_master_share_t * share = mmap (NULL, 4096, PROT_READ|PROT_WRITE, MAP_SHARED, sharefd, 0);